Move the element tree context manager to a common place
